Make it explicit that `IntersectionExclusion` uses an `ImmutableModuleExclusionSet`
... and optimize the latter for iteration: iterating on a set is expensive. Since this is an immutable
data structure, we can optimize iteration by using an array in addition to the set. So, calls to `contains`
are delegated to the set, but iterations are based on the array.

This change brings resolution time from 16s to 11s on a test case.
